More fixes to setup.sh

+++ setup.sh
@@ -1,5 +1,5 @@
-for zf in ("zshenv", "zprofile", "zshrc", "zprofile", "zlogin", "zlogout") do
-    if [ -f "${HOME}/.zsh/${zf}"]
-        ln -s ${HOME}/.zsh/${zf} ${HOME}/.${zf}
-    fi
+#!/bin/zsh
+for zf in "zshenv" "zprofile" "zshrc" "zprofile" "zlogin" "zlogout" 
+do
+	[[ -f "${HOME}/.zsh/${zf}" ]] && ln -s ${HOME}/.zsh/${zf} ${HOME}/.${zf}
 done
